% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm_extract.R
\name{emm_extract}
\alias{emm_extract}
\title{extraction of online surveys from EMM survey registry
Function to extract a selection of surveys and export an excel file}
\usage{
emm_extract(link)
}
\arguments{
\item{link}{a vector of characters containing urls}
}
\value{
function returns nothing but create an excel file
}
\description{
extraction of online surveys from EMM survey registry
Function to extract a selection of surveys and export an excel file
}
